requirements.txt
PyYAML==6.0.1
PyYAML==6.0.1
selenium==4.22.0
selenium_wire==5.1.0
